server/requirements.txt
amqp==2.5.2
billiard==3.6.3.0
blinker==1.4
celery==4.4.2
click==7.1.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==4.0.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==4.6.8
MarkupSafe==1.1.1
passlib==1.7.2
PyMySQL==0.9.3
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.2
six==1.14.0
SQLAlchemy==1.3.16
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
